k-ToM learning models
The aim of RecToMfunction is to code in a recursive, simpler more
general way the previous evolution1bisND, evolution2bisND etc.. that
were previously used as the evolution functions for a k-ToM. The aim is
to simplify the simulation and inversion of k-ToM player in the
following condition:
 2 persons – 2 actions game (given by a 2 (action player 1) x2 (action
player 2) x 2 (player) payoff matrix)
 the observation function for all the potential levels of the opponent
is assume to have the same form (i.e. same number of parameters) but
this need not to be the same as the one used by the focal player.
Example : we can simulate a focal player who is a 2-ToM with a bias that
assumes that the other player has no bias and is either a 1-ToM or 0-ToM
 Other Observation function (including perseveration …) would be easy
to implement but other evolution functions require change in
RecToMfunction (e.g. to allow a 1-ToM to assume that her opponent is
either a 0-ToM or an “environment”)
 If u is let empty, then useful to initialize in a consistent way the
hidden states (i.e. consistent f, df and uniform level).

Functions and scripts involved:
functions:
 demo_recur: simulate a game between a random opponent and a k-ToM
opponent (using RecToMfunction as evolution function and ObsRecGen as
observation function).
 RecToMfunction: main evolution function (recursive)
 ObsRecGen: observation function of a k-ToM player (important arguments
are hidden states, level, game and role). It can account for a bias if
the length of its parameter input is greater than 1.
 sizeXrec: compute the number of hidden states for a given level and
total number of parameter for the opponent
 defIndlev: create a structure with info about indexes of hidden states
(see below)
 sigmoid: sigmoidal transform (with a ceil to 10^-4)
 invsigmoid
 fplayer: computes the argument of the sigmoidal function in the
observation function as a function of the estimated probability of the
opponent playing option 1, the game matrix (2x2) and the temperature)
 bernoulli (p, N): function to draw following N Bernoulli variables of
parameter p.
Description of RecToMfunction
The main difficulty is to keep track of what each hidden state is
representing (this is done by the auxillliary function defindlev which
returns a structure containing the indexes of the hiddenstates
corresponding to the different potential levels of the opponent).
The general structure is as follow:
for a k=ToM player :
- k-1 first indexes : sufficient statistics for the level of the other
player (invsigmoid transform of the proba for the opponent to be of
level p=0,1,…, k-2, note that the proba to be of level k-1 can be
deduced )
- then, the simulated hidden states of the opponent for p=0, then p=1,…
p=k-1; . These indexes are kept in the structure indlev(p).X; (the size
is varying because of the recursive nature of the function)
- then f  and its partial derivatives following the parameters
(typically volatility and noise) of the opponent (I, the hypothesis he
is a p-ToM player). f is simply the inverse sigmoid of the probability
that the player will select an action at time t
- then the statistics (mean and variances) of the parameters of the
opponent (mean and variance for each parameter :[M1,V1,M2,V2,…])