corrected minor bug when estimating Volterra kernels on small-variance data
VBA_VolterraKernels.m used default priors on the noise precision when
etimating Volterra kernels. When the scale of the data was small, this
resulted in under-fitting of the Volterra decomposition. The correction
amounts to specify priorr on the noise precision, such that they match
the observed variance of the data.