minor changes for numerical robustness of k-ToM models
In rare circumstances, the estimation of k-ToM's opponents' parameters
may enter some form of limit cycle. This happen, e.g., when opponents'
decisions alternate with a fast pace due to random noise. The correction
applied here consists in bounding the numerical updates to the numerical
precision of sigmoid and inverse-sigmoid mappings.