corrected minor (numerical) issue with HGF learner
The HGF learning rule is based upon a one-step (non-iterated)
Gauss-Newton update of the posterior sufficients statistics. This
approximation can go awry for the third-level states, which would not
happen if an iterated Gauss-Newton scheme was used. The correction
simply regularizes the update in the rare event of negative precision of
the likelihood function.